feat: add models submodule
