+++ .flake8
@@ -1,2 +1,4 @@
 [flake8]
 max-line-length = 88
+
+ignore = F401,W503
