+++ mapo/models/__init__.py
@@ -1,0 +1,1 @@
+"""Collection of keras.Model factories."""
