+++ mapo/models/layers/__init__.py
@@ -1,0 +1,2 @@
+"""Collection of custom keras.layers."""
+from mapo.models.layers.action_squashing_layer import ActionSquashingLayer
